rename: 날씨 페이지 파일명 수정
related to: #26
